Improved role processing
Made the process more efficient so it iterates over less names as it
determines which roles to alter.
Additionally, it now attempts to be proactive in finding potential
issues when setting roles and skipping them if there are issues. This
had been causing issues with rate limiting due to having an enormous
number of failed requests.